+++ lib/open_graph_reader/configuration.rb
@@ -2,6 +2,8 @@ require 'singleton'
 
 module OpenGraphReader
   # The behavior of this library can be tweaked with some parameters.
+  # Note that configuration is global, changing it at runtime is not
+  # thread safe.
   #
   # @example
   #   OpenGraphReader.configure do |config|
@@ -10,7 +12,7 @@ module OpenGraphReader
   class Configuration
     include Singleton
 
-    # Strict mode (default: <tt>false</tt>)
+    # Strict mode (default: <tt>false</tt>).
     #
     # In strict mode, if the fetched site defines an unknown type
     # or property, {InvalidObjectError} is thrown instead of just ignoring
@@ -19,7 +21,7 @@ module OpenGraphReader
     # @return [Bool]
     attr_accessor :strict
 
-    # Validate required (default: <tt>true</tt>)
+    # Validate required (default: <tt>true</tt>).
     #
     # Validate that required properties exist. If this is enabled and
     # they do not,  {InvalidObjectError} is thrown.
@@ -27,7 +29,7 @@ module OpenGraphReader
     # @return [Bool]
     attr_accessor :validate_required
 
-    # Validate references (default: <tt>true</tt>)
+    # Validate references (default: <tt>true</tt>).
     #
     # If an object should be a reference to another object,
     # validate that it contains an URL. Be careful in turning this off,
@@ -36,7 +38,7 @@ module OpenGraphReader
     # @return [Bool]
     attr_accessor :validate_references
 
-    # Fallback to the title tag if og:title is missing (default: false)
+    # Fallback to the title tag if og:title is missing (default: <tt>false</tt>).
     #
     # The standard makes defining og:title required, but it's
     # a common practice to rely on the parser falling back to
@@ -44,6 +46,18 @@ module OpenGraphReader
     #
     # @return [Bool]
     attr_accessor :synthesize_title
+
+
+    # Guess image URL when it looks like a path (default: <tt>false</tt>).
+    #
+    # The standard requires og:image, og:image:url and og:image:secure_url
+    # to be a full URL. However it's common practice to put a path relative
+    # to the domain URL. When enabled, the library tries to guess the full
+    # URL from such a path. Note the object can still turn invalid if it fails
+    # to do so.
+    #
+    # @return [Bool]
+    attr_accessor :synthesize_image_url
 
     # @private
     def initialize
@@ -56,6 +70,7 @@ module OpenGraphReader
       @validate_required = true
       @validate_references = true
       @synthesize_title = false
+      @synthesize_image_url = false
     end
   end
 end