+++ lib/open_graph_reader/object/dsl/types.rb
@@ -1,4 +1,5 @@
 require 'date'
+require 'uri'
 
 require 'open_graph_reader/object/dsl'
 
@@ -10,22 +11,41 @@ module OpenGraphReader
         value.to_s
       end
 
-      # @see http://ogp.me/#url
-      define_type :url do |value, options|
-        value.to_s.tap {|value|
-          unless value.start_with?('http://') || value.start_with?('https://')
-             if options.has_key?(:to) && OpenGraphReader.config.validate_references
-              raise InvalidObjectError, "URL #{value.inspect} does not start with http:// or https://"
+      # @!method url(name, options={})
+      #   @option options [Bool] :image (false) Mark attribute as image to be eligible
+      #     for URL synthesization. See {Configuration#synthesize_image_url}.
+      #   @!macro define_type_description
+      #   @see http://ogp.me/#url
+      define_type_no_doc :url do |value, options|
+        value = value.to_s
+
+        unless value.start_with?('http://') || value.start_with?('https://')
+          if options[:image] && OpenGraphReader.config.synthesize_image_url
+            unless OpenGraphReader.current_origin
+              raise ArgumentError, "Enabled image url synthesization but didn't pass an origin"
             end
+
+            begin
+              value = "/#{value}" unless value.start_with? '/' # Normalize to absolute path
+              uri = URI.parse(OpenGraphReader.current_origin)
+              uri.path = value
+              value = uri.to_s
+            rescue
+              raise InvalidObjectError, "URL #{value.inspect} does not start with http:// or https:// and failed to synthesize a full URL"
+            end
+          elsif options.has_key?(:to) && OpenGraphReader.config.validate_references
+            raise InvalidObjectError, "URL #{value.inspect} does not start with http:// or https://"
           end
-        }
+        end
+
+        value
       end
 
       # @!method enum(name, allowed, options={})
       #   @param [Array<String>] allowed the list of allowed values
       #   @!macro define_type_description
-      # @see http://ogp.me/#enum
-      define_type_with_args :enum do |value, allowed|
+      #   @see http://ogp.me/#enum
+      define_type_no_doc :enum do |value, allowed|
         unless allowed.include? value
           raise InvalidObjectError, "Expected one of #{allowed.inspect} but was #{value.inspect}"
         end